+++ include/soloud_echofilter.h
@@ -37,7 +37,7 @@ namespace SoLoud
 		int mOffset;
 
 	public:
-		virtual void filter(float *aBuffer, int aSamples, int aChannels, float aSamplerate, float aTime);
+		virtual void filter(float *aBuffer, int aSamples, int aChannels, float aSamplerate, double aTime);
 		virtual ~EchoFilterInstance();
 		EchoFilterInstance(EchoFilter *aParent);
 	};
