+++ src/core/soloud_filter.cpp
@@ -70,7 +70,7 @@ namespace SoLoud
 		return 0;
 	}
 
-	void FilterInstance::updateParams(float aTime)
+	void FilterInstance::updateParams(double aTime)
 	{
 		int i;
 		for (i = 0; i < mNumParams; i++)
@@ -99,7 +99,7 @@ namespace SoLoud
 		mParamChanged |= 1 << aAttributeId;
 	}
 
-	void FilterInstance::fadeFilterParameter(int aAttributeId, float aTo, float aTime, float aStartTime)
+	void FilterInstance::fadeFilterParameter(int aAttributeId, float aTo, double aTime, double aStartTime)
 	{
 		if (aAttributeId < 0 || aAttributeId >= mNumParams || aTime <= 0 || aTo == mParam[aAttributeId])
 			return;
@@ -107,7 +107,7 @@ namespace SoLoud
 		mParamFader[aAttributeId].set(mParam[aAttributeId], aTo, aTime, aStartTime);
 	}
 
-	void FilterInstance::oscillateFilterParameter(int aAttributeId, float aFrom, float aTo, float aTime, float aStartTime)
+	void FilterInstance::oscillateFilterParameter(int aAttributeId, float aFrom, float aTo, double aTime, double aStartTime)
 	{
 		if (aAttributeId < 0 || aAttributeId >= mNumParams || aTime <= 0 || aFrom == aTo)
 			return;
@@ -123,7 +123,7 @@ namespace SoLoud
 		return mParam[aAttributeId];
 	}
 
-	void FilterInstance::filter(float *aBuffer, int aSamples, int aChannels, float aSamplerate, float aTime)
+	void FilterInstance::filter(float *aBuffer, int aSamples, int aChannels, float aSamplerate, double aTime)
 	{
 		int i;
 		for (i = 0; i < aChannels; i++)
@@ -132,7 +132,7 @@ namespace SoLoud
 		}
 	}
 
-	void FilterInstance::filterChannel(float *aBuffer, int aSamples, float aSamplerate, float aTime, int aChannel, int aChannels)
+	void FilterInstance::filterChannel(float *aBuffer, int aSamples, float aSamplerate, double aTime, int aChannel, int aChannels)
 	{
 	}
